gcc -Wextra and Makefile, fix 'show plan', create/unlink pid file

+++ src/board.c
@@ -917,7 +917,7 @@ void string_add(struct descriptor_data *d, char *str)
     }
 
     if (!(*d->str)) {
-	if (strlen(str) > d->max_str) {
+	if ((int)strlen(str) > d->max_str) {
 	    send_to_char("String too long - Truncated.\n\r",
 			 d->character);
 	    *(str + d->max_str) = '\0';
@@ -928,7 +928,7 @@ void string_add(struct descriptor_data *d, char *str)
 	strcpy(*d->str, str);
     }
     else {
-	if (strlen(str) + strlen(*d->str) > d->max_str) {
+	if ((int)(strlen(str) + strlen(*d->str)) > d->max_str) {
 	    send_to_char("String too long. Last line skipped.\n\r",
 			 d->character);
 	    terminator = 1;
@@ -1170,7 +1170,7 @@ void do_string(struct char_data *ch, char *arg, int cmd)
 
     if (*string) {	/* there was a string in the argument array */
 	/* NOTE: Max length of field 1 is length[1].. */
-	if (strlen(string) > length[field]) {
+	if ((int) strlen(string) > length[field]) {
 	    send_to_char("String too long - truncated.\n\r", ch);
 	    *(string + length[field]) = '\0';
 	}
